+++ packages/qiskit-algos-js/index.js
@@ -1,0 +1,23 @@
+/**
+ * @license
+ *
+ * Copyright (c) 2017-present, IBM Research.
+ *
+ * This source code is licensed under the Apache license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
+
+'use strict';
+
+const utils = require('@qiskit/utils');
+const qubits = require('jsqubits').jsqubits;
+
+const genHex = require('./lib/genHex');
+const { version } = require('./package');
+
+module.exports.version = version;
+
+module.exports.random = async opts => utils.genRandom(genHex, opts);
+
+// TODO: Make it fit with the doc.
+module.exports.factor = qubits.factor;