+++ packages/qiskit-algos-js/lib/genHex.js
@@ -1,0 +1,36 @@
+/**
+ * @license
+ *
+ * Copyright (c) 2017-present, IBM Research.
+ *
+ * This source code is licensed under the Apache license found in the
+ * LICENSE.txt file in the root directory of this source tree.
+ */
+
+'use strict';
+
+const utils = require('@qiskit/utils');
+const qubits = require('jsqubits').jsqubits;
+
+const { name } = require('../package');
+
+const dbg = utils.debug(name);
+
+module.exports = (len = 16) => {
+  let strHex = '';
+
+  // We make it in steps (char by char) to avoid a huge CPU load. It's not
+  // the best strategy for a cloud engine but this is a local one.
+  utils.times(len, () => {
+    // 1 char : 4 bits
+    const binChar = qubits('|0000>')
+      .hadamard(qubits.ALL)
+      .measure(qubits.ALL)
+      .asBitString();
+
+    strHex += utils.ayb.parseInt(binChar, 2, 16);
+  });
+
+  dbg('Random string (hex):', { strHex });
+  return strHex;
+};